pyproject.toml
[project]
name = "handoffagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "openai-agents>=0.3.3",
    "subprocess-tee>=0.4.2",
]
